+++ Source/SwiftLintBuiltInRules/Rules/Lint/UnneededOverrideRule.swift
@@ -89,8 +89,8 @@ private extension OverridableDecl {
             return false
         }
 
-        // Assume having @available changes behavior.
-        if attributes.contains(attributeNamed: "available") {
+        // Assume attributes change behavior.
+        guard attributes.isEmpty else {
             return false
         }