+++ CHANGELOG.md
@@ -22,6 +22,10 @@
   [SimplyDanny](https://github.com/SimplyDanny)
   [#5752](https://github.com/realm/SwiftLint/issues/5752)
 
+* Align left closure brace with associated parent function call in `contrasted_opening_brace` rule.  
+  [SimplyDanny](https://github.com/SimplyDanny)
+  [#5752](https://github.com/realm/SwiftLint/issues/5752)
+
 ## 0.56.1: Heat Pump Dryer
 
 #### Breaking