+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Main
+## 0.56.2: Heat Pump Dryer
 
 #### Breaking